remove sha checks from config
Since we are downloading the tarballs for the master branches of holonix en Holo-Host/nixpkgs, the hashes will change with every commit.
So as long as we do not check a specific version we have to omit this check. Otherwise nix-shell won't work after the next commit on one of the 2 projects.