Importing changes from solution branch (#16)
* Solution to the elements

* Reorganized

* Removed import

* Begin paths_exercise solution

* Solution working

* Fixed instructions

* Fixed paths exercise

* Fixed cargo lock

* Fixed bug

* adding solutino for entries exercise. Tests are in npm, so it's a real integration test

* fix test import

* removed sha checksum from nix config.nix
deleted zipped zome

* removed sha checksum from nix config.nix
deleted zipped zome

* remove sha checks from config

Since we are downloading the tarballs for the master branches of holonix en Holo-Host/nixpkgs, the hashes will change with every commit.
So as long as we do not check a specific version we have to omit this check. Otherwise nix-shell won't work after the next commit on one of the 2 projects.

* fix sha issue
remove and gitignore compiled zome
add convience script for running dna_util and npm tests at once from base folder

* Fixed gitignore

* Fixed solution

* gitignore dna's and vscode workspace files
correct README to use hc instead of dna-util
remove old workdir

* Adding solution of hashes exercise (#7)

* rename 1.elements folder to 3.elements

* hash exercise - not working

* Fixed nix

* updating to newer holochain version and hc-utils version;
change import hdk3 -> hdk

* hash exercise - work in progress - waiting to fix exercise after hdk & tryorama version updates

* merge nix file changes

* Hashes working!

* Fixed exercise

* Fix and further clarify the holochain gym link

The holochain gym link was directing to locahost so this fixes that.
Also increased the text the link covers to further clarify that the link is to the developer part of the holochain gym, not the whole gym.

* Align gym code with gym documentation

In the documentation for the gym exercise it says the line '#[hdk_extern()]' is not there and to add it. 
So both where it says what code is provided and what to do are not aligned with the actual code you get.
This pull requests fixes that by aligning the code provided with the text in the tutorial.

* fix readme
cleanup gitignore

* fix readme
add dna's to gitignore
delete old folders

* remove compiled dna from entries

* Update README.md

* remove unneeded comment

Co-authored-by: tixel <tim@pragmaticmind.com>
Co-authored-by: Guillem Cordoba <guillem.cordoba@gmail.com>
Co-authored-by: Stephen B Purkiss <me@stephenbpurkiss.com>
Co-authored-by: Guillem Córdoba <guillemcordoba@users.noreply.github.com>

* fix link in readme

* change readme text

* Fixed solution

* Attempt CI

* Attempt CI

* Added run all tests script

* Added x permission

* Removed hc_utils deps

* Fixed paths exercise

* Upgraded to latest core deps

* add solutions for remote call & capability token exercises (#13)

the CI build failed and was cancelled after 360 minutes. But it does not look related to the exercises themself. So I merged everything.

* attempt 2: removing more hc-utils deps && trying dumb CI script (#15)

* add headers test
reverted version upgrade of core dependencies

* remove hc-utils dependencies

* trying out dumb script to make CI on github work

* removing tests for basic/3.elements

* fix script to run cargo test on 3.element instead integration test
delete compiled dna's
ignore failing test for 3.remote-call -> should become test that validate if async call throwns error.

Co-authored-by: tixel <tim@pragmaticmind.com>

* adding headers exercise

Co-authored-by: Guillem Cordoba <guillem.cordoba@gmail.com>
Co-authored-by: tixel <tim@pragmaticmind.com>
Co-authored-by: Stephen B Purkiss <me@stephenbpurkiss.com>
Co-authored-by: Guillem Córdoba <guillemcordoba@users.noreply.github.com>
Co-authored-by: Joshua Parkin <joshuadparkin@gmail.com>